2048 in c++
The objective of the game is to get the number 2048 using additions of
the number two and its multiples.

You will have a grid of 16 tiles. Two numbers will be given: usually two
number twos, maybe number four.

Move up or down, left or right using w,s,a, and d trying to join two
equal numbers.

When two equal numbers are in touch, they will add up.

If we are run out of equal numbers on our grid, or we can not put them
in touch, the game will provide us another two, a four or even an eight
so we can keep on playing. If there are no free tiles on our grid, the
game ends.

By adding numbers, we get higher numbers and we can approach to 2048,
which is the goal of the game.